+++ src/main/java/ie/wellbeing/service/MembershipContextService.java
@@ -3,6 +3,6 @@ package ie.wellbeing.service;
 
 public interface MembershipContextService {
 
-    public void changeStateTo(MembershipState newState, Integer userId);
+    public void changeStateTo(MembershipState newState, Long userId);
     public String handle();
 }
