+++ src/main/java/ie/wellbeing/service/MembershipService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface MembershipService {
     String createMembership(MembershipRequest membershipRequest, String siteURL) throws Exception;
 
-    void updateMembershipDetails(Integer userId, String type);
+    void updateMembershipDetails(Long userId, String type);
 
     List<MembershipDetails> getAllUsersMembershipDetails();
 }
